Reef-Server/pyproject.toml
[project]
name = "reef-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.8",
    "gunicorn>=23",
    "httpx>=0.28",
    "openai>=1.60",
    "PyJWT[crypto]>=2.10",
    "PyMuPDF>=1.25",
    "numpy>=1.26",
    "rdp>=0.8",
    "circle-fit>=0.2",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
]
